+++ src/main/webapp/index.jsp
@@ -8,7 +8,7 @@
 %>
 		<div class="row">
 			<div class="col-12" align="center">
-				<div class="alert alert-primary" role="alert">
+				<div class="alert alert-success" role="alert">
 				  	<%= message %>
 				</div>
 			</div>
